requirements.txt
amqp==5.3.1
asgiref==3.8.1
beautifulsoup4==4.13.4
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
diff-match-patch==20241021
Django==5.1
django-allauth==65.9.0
django-bootstrap5==25.1
django-celery-beat==2.8.1
django-crispy-forms==2.4
django-import-export==4.3.7
django-ranged-response==0.2.0
django-simple-captcha==0.6.2
django-timezone-field==7.1
django-tinymce==4.1.0
django_celery_results==2.6.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.10
kombu==5.5.4
lxml==6.0.0
numpy==2.2.5
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.3.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
redis==5.2.1
requests==2.32.4
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
